minor refactoring and enhancement
