+++ rl4co/models/nn/ops.py
@@ -1,7 +1,7 @@
-import math
+from typing import Tuple, Union
 
-import torch.nn as nn
 import torch
+import torch.nn as nn
 
 
 class SkipConnection(nn.Module):
@@ -13,24 +13,37 @@ class SkipConnection(nn.Module):
         return x + self.module(x)
 
 
+class AdaptiveSequential(nn.Sequential):
+    def forward(
+        self, *inputs: Union[Tuple[torch.Tensor], torch.Tensor]
+    ) -> Union[Tuple[torch.Tensor], torch.Tensor]:
+        for module in self._modules.values():
+            if type(inputs) == tuple:
+                inputs = module(*inputs)
+            else:
+                inputs = module(inputs)
+        return inputs
+
+
 class Normalization(nn.Module):
     def __init__(self, embed_dim, normalization="batch"):
         super(Normalization, self).__init__()
-        if normalization != 'layer':
-            normalizer_class = {"batch": nn.BatchNorm1d, "instance": nn.InstanceNorm1d}.get(
-                normalization, None
-            )
-    
+        if normalization != "layer":
+            normalizer_class = {
+                "batch": nn.BatchNorm1d,
+                "instance": nn.InstanceNorm1d,
+            }.get(normalization, None)
+
             self.normalizer = normalizer_class(embed_dim, affine=True)
         else:
-            self.normalizer = 'layer'
-            
+            self.normalizer = "layer"
+
     def forward(self, x):
         if isinstance(self.normalizer, nn.BatchNorm1d):
             return self.normalizer(x.view(-1, x.size(-1))).view(*x.size())
         elif isinstance(self.normalizer, nn.InstanceNorm1d):
             return self.normalizer(x.permute(0, 2, 1)).permute(0, 2, 1)
-        elif self.normalizer == 'layer':
+        elif self.normalizer == "layer":
             return (x - x.mean((1, 2)).view(-1, 1, 1)) / torch.sqrt(
                 x.var((1, 2)).view(-1, 1, 1) + 1e-05
             )
